minor fixes
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code@6754 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ PalladioCM.CodeGen/trunk/de.uka.ipd.sdq.simucomframework/src/de/uka/ipd/sdq/simucomframework/SimuComConfig.java
@@ -45,7 +45,7 @@ public class SimuComConfig implements Serializable {
 					MAXIMUM_MEASUREMENT_COUNT));
 			this.verboseLogging = (Boolean)configuration.get(
 					VERBOSE_LOGGING);
-			this.datasourceID = (Long)configuration.get(
+			this.datasourceID = (Integer)configuration.get(
 					DATASOURCE_ID);
 			this.runNumber = runNo;
 			this.isDebug = debug;
